+++ src/confluent_kafka/src/Admin.c
@@ -492,17 +492,17 @@ static PyObject *Admin_create_topics (Handle *self, PyObject *args,
                                 goto err;
                         }
 
-                        int t;
+                        int temp_partitions;
                         if (newt->num_partitions == -1) {
-                                t = PyList_Size(newt->replica_assignment);
+                                temp_partitions = PyList_Size(newt->replica_assignment);
                         } else {
-                                t = newt->num_partitions;
+                                temp_partitions = newt->num_partitions;
                         }
                         if (!Admin_set_replica_assignment(
                                     "CreateTopics", (void *)c_objs[i],
                                     newt->replica_assignment,
-                                    t,
-                                    t, 
+                                    temp_partitions,
+                                    temp_partitions, 
                                     "num_partitions")) {
                                 i++;
                                 goto err;